fix(files): prevent right panel reload flicker
Made-with: Proma
